Splitting Fixed DayCalendar to its own file and tests
